requirements.txt
python-telegram-bot==13.11
environs==9.5.0
redis==4.2.0
key-value-db==1.1.7